remove engine.h from comp.hpp
